Move RedirectTargetProvider in new namespace
